Fix namespace objects' creation. This resovled 2 failures and an error of test_*_splode_* in xml/test_node_set.rb.
However, weird 4 erros have started reporting from test_relax_ng.rb by rake test. All tests have passed when test_relax_ng.rb gets run from test_jruby.rb, but 4 of them report errors from rake test. Perhaps, JRuby caches some sort of states that cause errors. This is really hard to debug since test_relax_ng.rb itself doesn't have any problem. Besides, so many tests run before test_relax_ng.rb in rake test, which makes it hard to figure out exactly what test is the culprit.

+++ ext/java/nokogiri/XmlNodeSet.java
@@ -1,7 +1,9 @@
 package nokogiri;
 
 import java.util.List;
+
 import nokogiri.internals.NokogiriHelpers;
+
 import org.jruby.Ruby;
 import org.jruby.RubyArray;
 import org.jruby.RubyClass;
@@ -55,10 +57,12 @@ public class XmlNodeSet extends RubyObject {
     }
 
     public void relink_namespace(ThreadContext context) {
-        List n = this.nodes.getList();
+        List<?> n = this.nodes.getList();
 
-        for(int i = 0; i < n.size(); i++) {
-            ((XmlNode) n.get(i)).relink_namespace(context);
+        for (int i = 0; i < n.size(); i++) {
+            if (n.get(i) instanceof XmlNode) {
+                ((XmlNode) n.get(i)).relink_namespace(context);
+            }
         }
     }
 
@@ -131,8 +135,10 @@ public class XmlNodeSet extends RubyObject {
     public IRubyObject unlink(ThreadContext context){
         IRubyObject[] arr = this.nodes.toJavaArrayUnsafe();
         long length = arr.length;
-        for(int i = 0; i < length; i++) {
-            ((XmlNode) arr[i] ).unlink(context);
+        for (int i = 0; i < length; i++) {
+            if (arr[i] instanceof XmlNode) {
+                ((XmlNode) arr[i] ).unlink(context);
+            }
         }
         return this;
     }
